TPC: fix compiler warnings (#571)
* Fix compiler warning

fixes warning: comparison of unsigned expression < 0 is always false [-Wtautological-compare]

* Fix compiler warning

warning: control reaches end of non-void function [-Wreturn-type]

* Fix compiler warning

warning: format specifies type 'double' but the argument has type 'short' [-Wformat]

* Fix compiler warning: operator precedence

* Fix compiler warning

warning: control reaches end of non-void function

* Fix nonvirtual destructor

An abstract class with purely virtual functions must provide
a virtual destructor

* TPC: virtual destructor

* TPC: fix inconsistent override

+++ Detectors/TPC/base/src/Painter.cxx
@@ -176,7 +176,7 @@ TH2* Painter::getHistogram2D(const CalArray<T>& calArray)
       const auto val = calArray.getValue(pad);
       const int cpad = ipad - padsInRow/2;
       hist->Fill(irow, cpad, val);
-      printf("%d %d: %f\n", irow, cpad, val);
+      printf("%d %d: %f\n", irow, cpad, (double)val);
     }
   }
   return hist;
